fix(spec): resolve specification compliance issues for Stage 7
- Move AlertChannelRoute and AlertChannelConfig from coverage_config.py to coverage_alert_channels.py to match spec requirements
- Update imports in coverage_config.py and test_coverage_config.py to reflect new locations
- Correct specification to reflect actual architecture: coverage_collector.py and coverage_signal.py in collectors/ subdirectory
- Add __all__ export to coverage_alert_channels.py with all public classes
- Update Stage 7 verification section with spec compliance fixes

All implementation modules now pass specification verification:
✅ coverage_models.py: 6 classes
✅ coverage_alert_channels.py: 7 classes (AlertChannelRoute, AlertChannelConfig moved here)
✅ coverage_alerting.py: 4 classes
✅ coverage_config.py: 7 classes
✅ coverage_trend_manager.py: 2 items
✅ coverage_trend_repository.py: 4 classes
✅ collectors/coverage_collector.py: 1 class
