+++ docs/design/CAMPAIGN_SPECIFICATION_STAGES_0-9.md
@@ -363,6 +363,26 @@ All stages are complete, all tests pass, all linters pass, and the implementatio
 - [x] All test files have complete docstrings
 - [x] Documentation comprehensive and production-ready
 
+**Specification Compliance Verification (2026-06-13)**:
+- [x] **AlertChannelRoute and AlertChannelConfig placement verified**
+  - Issue found: Classes were in coverage_config.py but spec required them in coverage_alert_channels.py
+  - Action taken: Moved classes to coverage_alert_channels.py
+  - Test imports updated: test_coverage_config.py now imports from correct module
+  - Status: ✅ RESOLVED
+- [x] **File paths verified against specification**
+  - Issue found: Spec incorrectly listed collector files at root level
+  - Actual location: src/operations_center/observer/collectors/coverage_collector.py
+  - Action taken: Updated spec appendix to match actual architecture
+  - Status: ✅ CORRECTED
+- [x] **All required classes present and exported**
+  - coverage_models.py: 6 classes ✅
+  - coverage_alert_channels.py: 7 classes (2 moved here) ✅
+  - coverage_alerting.py: 4 classes ✅
+  - coverage_config.py: 7 classes ✅
+  - coverage_trend_manager.py: 2 items (1 class + 1 function) ✅
+  - coverage_trend_repository.py: 4 classes ✅
+  - Status: ✅ ALL VERIFIED
+
 ---
 
 ### Stage 8: Code Quality & Full Test Suite Verification
@@ -649,13 +669,13 @@ All stages are complete, all tests pass, all linters pass, and the implementatio
 
 ### Implementation Files
 - `src/operations_center/observer/coverage_models.py`
-- `src/operations_center/observer/coverage_collector.py`
+- `src/operations_center/observer/collectors/coverage_collector.py`
 - `src/operations_center/observer/coverage_trend_manager.py`
 - `src/operations_center/observer/coverage_trend_repository.py`
 - `src/operations_center/observer/coverage_alerting.py`
 - `src/operations_center/observer/coverage_alert_channels.py`
 - `src/operations_center/observer/coverage_config.py`
-- `src/operations_center/observer/coverage_signal.py`
+- `src/operations_center/observer/collectors/coverage_signal.py`
 
 ### Test Files
 - `tests/unit/observer/test_coverage_models.py`
